[PRMT-4566] Rename test files
